WIP POC Generate Rust code from the tutorial text
This updates the Asciidoctor documents to allow generating the final
source files at the end of each chapter directly from the tutorial
text.

The inline source blocks in the docs now contain tags delimiting
sections of the code we want to pull into the final files.

We're adding separate documents that can include these tagged regions
and output the final `*.rs` files identical to the hardcoded ones in
the `src/bin` directories.

There are several benefits of generating these Rust files out of the
document itself:

1. The docs & corresponding sources are always in sync (the doc being
the single source of truth)
2. The Rust sources are generating pretty in a similar way as a person
going over the tutorial would
3. There is reduced code duplication -- if files 3 to 13 contain the
same function and we want to change it, there's only one place to do
that (right now one would have to change it in all 10 Rust files plus the
tutorial doc)

Downsides:

1. The files are no longer checked out in the repo so you can't clone
it and run the examples. We could still generate & check them after the fact, but
that might confuse potential contributors (they could try & fix those
files without realising they were autogenerated)
2. You can't easily fix stuff with tools such as `cargo fmt`
automatically any more -- they don't work on adoc which is the single
source of truth. We could generate RS and run it in the CI (we should
verify they compile anyway) and fail if `cargo fmt` would make any
changes.

Unfortunately, the comments indicating the source tags (regions) are
visible in the HTML generated from adoc. This is clearly undesirable
and I could not figure out how to hide those directly.

The current workaround is: add `[role=hide]` to the "source-of-truth"
code block (no need to annotate it with `[source,rust]` any more) --
that will apply the `hide` CSS class to the block, making it invisible
thanks to the included stylesheet.

And then immediately follow it up with another (visible) Rust source block
`include::file.adoc[tag=...]` which adds the whole thing and formats it
as Rust. The `include` process does strip all tags.

This is not great -- it adds a bunch of boiler plate, but it seems to
work fine. I'll try to see if there either is a way to hide the tags
in asciidoctor or if that's something that could be added. But in the
meantime, I think this is a decent workaround (the benefits are imho
awesome).