Move `lol.adoc` to `doc/output-templates/`
The general idea is: can we keep the canonical adoc sources in the
`/doc/` directory and then move both the .rs templates as well as the
final adoc bits (e.g. the single include that filters out tags)
elsewhere?

So the user would always see the canonical sources first -- those are
the files most likely to be edited -- and then dig into the templates
only if need be.