Build "part 1 graphics" file from output-template
This updates the `Makefile` to only build code that's in the output
templates file. It includes a file from a directory above (the
"source" adoc) and produces an identical output HTML page with the
tags stripped out.

I wish we didn't have to basically create these single-line
include-only adoc files, but it matches the source code ones and it
gives us everything we need.