Split `output-templates` to `adoc` and `rs` dirs
Some documents, such as the fifth part use the same name for the
tutorial and source listing.

And even though the output will be either html or rs, both "include
template docs" have the same `part-5-combat.adoc` name.

So this splits the adoc includes and rust templates into separate
folders and updates the Makefile to read both.